project_app/requirements.txt
Flask==3.0.3
numpy==1.26.4
pillow==10.4.0
tensorflow==2.16.2
Werkzeug==3.0.3
